+++ crates/fuel-core/src/database/coin.rs
@@ -1,14 +1,13 @@
 use crate::{
     database::{
-        database_description::off_chain::OffChain,
-        Database,
+        OffChainIterableKeyValueView,
+        OnChainIterableKeyValueView,
     },
     fuel_core_graphql_api::storage::coins::{
         owner_coin_id_key,
         OwnedCoins,
     },
 };
-use fuel_core_chain_config::TableEntry;
 use fuel_core_storage::{
     iter::{
         IterDirection,
@@ -27,9 +26,8 @@ use fuel_core_types::{
         UtxoId,
     },
 };
-use itertools::Itertools;
 
-impl Database<OffChain> {
+impl OffChainIterableKeyValueView {
     pub fn owned_coins_ids(
         &self,
         owner: &Address,
@@ -55,7 +53,7 @@ impl Database<OffChain> {
     }
 }
 
-impl Database {
+impl OnChainIterableKeyValueView {
     pub fn coin(&self, utxo_id: &UtxoId) -> StorageResult<CompressedCoin> {
         let coin = self
             .storage_as_ref::<Coins>()
@@ -64,12 +62,5 @@ impl Database {
             .into_owned();
 
         Ok(coin)
-    }
-
-    pub fn iter_coins(
-        &self,
-    ) -> impl Iterator<Item = StorageResult<TableEntry<Coins>>> + '_ {
-        self.iter_all::<Coins>(None)
-            .map_ok(|(key, value)| TableEntry { key, value })
     }
 }