+++ crates/fuel-core/src/database/storage.rs
@@ -1,77 +1,35 @@
-use crate::database::{
-    database_description::DatabaseDescription,
-    Database,
-};
+use crate::state::generic_database::GenericDatabase;
 use fuel_core_storage::{
     structured_storage::StructuredStorage,
+    transactional::{
+        ConflictPolicy,
+        Modifiable,
+        StorageTransaction,
+    },
     Error as StorageError,
     Mappable,
-    MerkleRoot,
-    MerkleRootStorage,
     Result as StorageResult,
-    StorageAsRef,
-    StorageInspect,
-    StorageRead,
-    StorageSize,
-};
-use std::borrow::Cow;
-
-#[cfg(feature = "test-helpers")]
-use fuel_core_storage::transactional::{
-    ConflictPolicy,
-    Modifiable,
-    StorageTransaction,
-};
-#[cfg(feature = "test-helpers")]
-use fuel_core_storage::{
     StorageAsMut,
     StorageBatchMutate,
+    StorageInspect,
     StorageMutate,
     StorageWrite,
 };
 
-impl<Description, Stage, M> StorageInspect<M> for Database<Description, Stage>
+impl<Storage, M> StorageMutate<M> for GenericDatabase<Storage>
 where
-    Description: DatabaseDescription,
     M: Mappable,
-    for<'a> StructuredStorage<&'a Self>: StorageInspect<M, Error = StorageError>,
-{
-    type Error = StorageError;
-
-    fn get(&self, key: &M::Key) -> StorageResult<Option<Cow<M::OwnedValue>>> {
-        let storage = StructuredStorage::new(self);
-        let value = storage.storage::<M>().get(key)?;
-
-        if let Some(cow) = value {
-            Ok(Some(Cow::Owned(cow.into_owned())))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn contains_key(&self, key: &M::Key) -> StorageResult<bool> {
-        StructuredStorage::new(self)
-            .storage::<M>()
-            .contains_key(key)
-    }
-}
-
-#[cfg(feature = "test-helpers")]
-impl<Description, Stage, M> StorageMutate<M> for Database<Description, Stage>
-where
-    Description: DatabaseDescription,
-    M: Mappable,
-    for<'a> StructuredStorage<&'a Self>: StorageInspect<M, Error = StorageError>,
-    for<'a> StorageTransaction<&'a Self>: StorageMutate<M, Error = StorageError>,
     Self: Modifiable,
+    StructuredStorage<Storage>: StorageInspect<M, Error = StorageError>,
+    for<'a> StorageTransaction<&'a Storage>: StorageMutate<M, Error = StorageError>,
 {
     fn insert(
         &mut self,
         key: &M::Key,
         value: &M::Value,
-    ) -> StorageResult<Option<M::OwnedValue>> {
+    ) -> Result<Option<M::OwnedValue>, Self::Error> {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );
@@ -80,9 +38,9 @@ where
         Ok(prev)
     }
 
-    fn remove(&mut self, key: &M::Key) -> StorageResult<Option<M::OwnedValue>> {
+    fn remove(&mut self, key: &M::Key) -> Result<Option<M::OwnedValue>, Self::Error> {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );
@@ -92,56 +50,16 @@ where
     }
 }
 
-impl<Description, Stage, M> StorageSize<M> for Database<Description, Stage>
+impl<Storage, M> StorageWrite<M> for GenericDatabase<Storage>
 where
-    Description: DatabaseDescription,
     M: Mappable,
-    for<'a> StructuredStorage<&'a Self>: StorageSize<M, Error = StorageError>,
-{
-    fn size_of_value(&self, key: &M::Key) -> StorageResult<Option<usize>> {
-        <_ as StorageSize<M>>::size_of_value(&StructuredStorage::new(self), key)
-    }
-}
-
-impl<Description, Stage, Key, M> MerkleRootStorage<Key, M>
-    for Database<Description, Stage>
-where
-    Description: DatabaseDescription,
-    M: Mappable,
-    for<'a> StructuredStorage<&'a Self>: MerkleRootStorage<Key, M, Error = StorageError>,
-{
-    fn root(&self, key: &Key) -> StorageResult<MerkleRoot> {
-        StructuredStorage::new(self).storage::<M>().root(key)
-    }
-}
-
-impl<Description, Stage, M> StorageRead<M> for Database<Description, Stage>
-where
-    Description: DatabaseDescription,
-    M: Mappable,
-    for<'a> StructuredStorage<&'a Self>: StorageRead<M, Error = StorageError>,
-{
-    fn read(&self, key: &M::Key, buf: &mut [u8]) -> StorageResult<Option<usize>> {
-        StructuredStorage::new(self).storage::<M>().read(key, buf)
-    }
-
-    fn read_alloc(&self, key: &M::Key) -> StorageResult<Option<Vec<u8>>> {
-        StructuredStorage::new(self).storage::<M>().read_alloc(key)
-    }
-}
-
-#[cfg(feature = "test-helpers")]
-impl<Description, Stage, M> StorageWrite<M> for Database<Description, Stage>
-where
-    Description: DatabaseDescription,
-    M: Mappable,
-    for<'a> StructuredStorage<&'a Self>: StorageInspect<M, Error = StorageError>,
-    for<'a> StorageTransaction<&'a Self>: StorageWrite<M, Error = StorageError>,
+    StructuredStorage<Storage>: StorageInspect<M, Error = StorageError>,
+    for<'a> StorageTransaction<&'a Storage>: StorageWrite<M, Error = StorageError>,
     Self: Modifiable,
 {
     fn write(&mut self, key: &M::Key, buf: &[u8]) -> Result<usize, Self::Error> {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );
@@ -156,7 +74,7 @@ where
         buf: &[u8],
     ) -> Result<(usize, Option<Vec<u8>>), Self::Error> {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );
@@ -167,7 +85,7 @@ where
 
     fn take(&mut self, key: &M::Key) -> Result<Option<Vec<u8>>, Self::Error> {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );
@@ -177,13 +95,11 @@ where
     }
 }
 
-#[cfg(feature = "test-helpers")]
-impl<Description, Stage, M> StorageBatchMutate<M> for Database<Description, Stage>
+impl<Storage, M> StorageBatchMutate<M> for GenericDatabase<Storage>
 where
-    Description: DatabaseDescription,
     M: Mappable,
-    for<'a> StructuredStorage<&'a Self>: StorageInspect<M, Error = StorageError>,
-    for<'a> StorageTransaction<&'a Self>: StorageBatchMutate<M, Error = StorageError>,
+    StructuredStorage<Storage>: StorageInspect<M, Error = StorageError>,
+    for<'a> StorageTransaction<&'a Storage>: StorageBatchMutate<M, Error = StorageError>,
     Self: Modifiable,
 {
     fn init_storage<'a, Iter>(&mut self, set: Iter) -> StorageResult<()>
@@ -193,7 +109,7 @@ where
         M::Value: 'a,
     {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );
@@ -209,7 +125,7 @@ where
         M::Value: 'a,
     {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );
@@ -224,7 +140,7 @@ where
         M::Key: 'a,
     {
         let mut transaction = StorageTransaction::transaction(
-            &*self,
+            self.as_ref(),
             ConflictPolicy::Overwrite,
             Default::default(),
         );